task-driver, common: Refactor descriptors into common crate
